Rename file and split python files
