+++ src/workers/python/python_package.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
-
+OUT_DIR=./python_package
+SRC_DIR=./papyros
 set -eux
 
-rm -rf ./package ./package.tar
-mkdir ./package
-pip install -t ./package python-runner friendly_traceback jedi
-cp -r ./papyros ./package
-cd ./package
-tar cfv ./../package.tar *
+rm -rf $OUT_DIR $OUT_DIR.tar.load_by_url
+mkdir $OUT_DIR
+pip install -t $OUT_DIR python-runner friendly_traceback jedi
+cp -r $SRC_DIR $OUT_DIR
+cd $OUT_DIR
+tar cfv ./.$OUT_DIR.tar.load_by_url *
